agents/weather_agent/requirements.txt
fastapi
google-adk
google-generativeai
pydantic